backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
pandas-ta
requests
feedparser
textblob
python-dotenv
scikit-learn
diskcache
cachetools
gunicorn
